add test and move some around
