+++ test/master-playlist-controller.test.js
@@ -9,10 +9,6 @@ import {
 } from './test-helpers.js';
 import manifests from './test-manifests.js';
 import { MasterPlaylistController } from '../src/master-playlist-controller';
-import {
-  mimeTypesForPlaylist_,
-  mapLegacyAvcCodecs_
-} from '../src/util/codecs';
 /* eslint-disable no-unused-vars */
 // we need this so that it can register hls with videojs
 import { Hls } from '../src/videojs-http-streaming';
@@ -21,50 +17,6 @@ import Playlist from '../src/playlist';
 import Config from '../src/config';
 import PlaylistLoader from '../src/playlist-loader';
 import DashPlaylistLoader from '../src/dash-playlist-loader';
-
-const generateMedia = function(isMaat, isMuxed, hasVideoCodec, hasAudioCodec, isFMP4) {
-  const codec = (hasVideoCodec ? 'avc1.deadbeef' : '') +
-    (hasVideoCodec && hasAudioCodec ? ',' : '') +
-    (hasAudioCodec ? 'mp4a.40.E' : '');
-  const master = {
-    mediaGroups: {},
-    playlists: []
-  };
-  const media = {
-    attributes: {}
-  };
-
-  if (isMaat) {
-    master.mediaGroups.AUDIO = {
-      test: {
-        demuxed: {
-          uri: 'foo.bar'
-        }
-      }
-    };
-
-    if (isMuxed) {
-      master.mediaGroups.AUDIO.test.muxed = {};
-    }
-    media.attributes.AUDIO = 'test';
-  }
-
-  if (isFMP4) {
-    // This is not a great way to signal that the playlist is fmp4 but
-    // this is how we currently detect it in HLS so let's emulate it here
-    media.segments = [
-      {
-        map: 'test'
-      }
-    ];
-  }
-
-  if (hasVideoCodec || hasAudioCodec) {
-    media.attributes.CODECS = codec;
-  }
-
-  return [master, media];
-};
 
 QUnit.module('MasterPlaylistController', {
   beforeEach(assert) {
@@ -858,7 +810,7 @@ function(assert) {
   assert.equal(this.player.tech_.hls.stats.bandwidth, 1, 'bandwidth we set above');
 });
 
-QUnit.test('blacklists switching between playlists with incompatible audio codecs',
+QUnit.test('does not blacklist switching between playlists with different audio profiles',
 function(assert) {
   let alternatePlaylist;
 
@@ -883,9 +835,32 @@ function(assert) {
     this.masterPlaylistController.masterPlaylistLoader_.master.playlists[1];
   assert.equal(alternatePlaylist.excludeUntil,
                undefined,
-               'not excluded incompatible playlist');
+               'not excluded playlist');
   // verify stats
   assert.equal(this.player.tech_.hls.stats.bandwidth, 1, 'bandwidth we set above');
+});
+
+QUnit.test('blacklists playlists with unsupported codecs before initial selection',
+function(assert) {
+  this.masterPlaylistController.selectPlaylist = () => {
+    assert.equal(
+      this.masterPlaylistController.master().playlists[0].excludeUntil,
+      Infinity,
+      'Blacklists unsupported playlist before initial selection');
+  }
+
+  openMediaSource(this.player, this.clock);
+
+  // master
+  this.requests.shift().respond(200, null,
+    '#EXTM3U\n' +
+    '#EXT-X-STREAM-INF:BANDWIDTH=1,CODECS="unsupporte.dc0dec,mp4a.40.5"\n' +
+    'media.m3u8\n' +
+    '#EXT-X-STREAM-INF:BANDWIDTH=10000,CODECS="avc1.4d400d,mp4a.40.2"\n' +
+    'media1.m3u8\n');
+
+  // media
+  this.standardXHRResponse(this.requests.shift());
 });
 
 QUnit.test('updates the combined segment loader on media changes', function(assert) {
@@ -2395,216 +2370,4 @@ QUnit.test('properly configures loader mime types', function(assert) {
                'correct mime type for audio segment loader');
   assert.ok(audioMimeTypeCalls[0][1] instanceof videojs.EventTarget,
             'passed a source buffer emitter to audio segment loader');
-});
-
-QUnit.module('Codec to MIME Type Conversion');
-
-const testMimeTypes = function(assert, isFMP4) {
-  let container = isFMP4 ? 'mp4' : 'mp2t';
-
-  let videoMime = `video/${container}`;
-  let audioMime = `audio/${container}`;
-
-  // no MAAT
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(false, true, false, false, isFMP4)),
-    [`${videoMime}; codecs="avc1.4d400d, mp4a.40.2"`],
-    `no MAAT, container: ${container}, codecs: none`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(false, true, true, false, isFMP4)),
-    [`${videoMime}; codecs="avc1.deadbeef"`],
-    `no MAAT, container: ${container}, codecs: video`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(false, true, false, true, isFMP4)),
-    [`${audioMime}; codecs="mp4a.40.E"`],
-    `no MAAT, container: ${container}, codecs: audio`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(false, true, true, true, isFMP4)),
-    [`${videoMime}; codecs="avc1.deadbeef, mp4a.40.E"`],
-    `no MAAT, container: ${container}, codecs: video, audio`);
-
-  // MAAT, not muxed
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(true, false, false, false, isFMP4)),
-    [`${videoMime}; codecs="avc1.4d400d"`,
-     `${audioMime}; codecs="mp4a.40.2"`],
-    `MAAT, demuxed, container: ${container}, codecs: none`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(true, false, true, false, isFMP4)),
-    [`${videoMime}; codecs="avc1.deadbeef"`,
-     `${audioMime}; codecs="mp4a.40.2"`],
-    `MAAT, demuxed, container: ${container}, codecs: video`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(true, false, false, true, isFMP4)),
-    [`${videoMime}; codecs="mp4a.40.E"`,
-     `${audioMime}; codecs="mp4a.40.E"`],
-    `MAAT, demuxed, container: ${container}, codecs: audio`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(true, false, true, true, isFMP4)),
-    [`${videoMime}; codecs="avc1.deadbeef"`,
-     `${audioMime}; codecs="mp4a.40.E"`],
-    `MAAT, demuxed, container: ${container}, codecs: video, audio`);
-
-  // MAAT, muxed
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(true, true, false, false, isFMP4)),
-    [`${videoMime}; codecs="avc1.4d400d, mp4a.40.2"`,
-     `${audioMime}; codecs="mp4a.40.2"`],
-    `MAAT, muxed, container: ${container}, codecs: none`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(true, true, true, false, isFMP4)),
-    [`${videoMime}; codecs="avc1.deadbeef, mp4a.40.2"`,
-     `${audioMime}; codecs="mp4a.40.2"`],
-    `MAAT, muxed, container: ${container}, codecs: video`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(true, true, false, true, isFMP4)),
-    [`${videoMime}; codecs="mp4a.40.E"`,
-     `${audioMime}; codecs="mp4a.40.E"`],
-    `MAAT, muxed, container: ${container}, codecs: audio`);
-
-  assert.deepEqual(mimeTypesForPlaylist_.apply(null,
-      generateMedia(true, true, true, true, isFMP4)),
-    [`${videoMime}; codecs="avc1.deadbeef, mp4a.40.E"`,
-     `${audioMime}; codecs="mp4a.40.E"`],
-    `MAAT, muxed, container: ${container}, codecs: video, audio`);
-};
-
-QUnit.test('recognizes muxed codec configurations', function(assert) {
-  testMimeTypes(assert, false);
-  testMimeTypes(assert, true);
-});
-
-// dash audio playlist won't have a URI but will have resolved playlists
-QUnit.test('content demuxed if alt audio URI not present but playlists present',
-function(assert) {
-  const media = {
-    attributes: {
-      AUDIO: 'test',
-      CODECS: 'avc1.deadbeef, mp4a.40.E'
-    },
-    segments: [
-      // signal fmp4
-      { map: 'test' }
-    ]
-  };
-  const master = {
-    mediaGroups: {
-      AUDIO: {
-        test: {
-          demuxed: {
-            uri: 'foo.bar'
-          }
-        }
-      }
-    },
-    playlists: [media]
-  };
-
-  assert.deepEqual(mimeTypesForPlaylist_(master, media),
-                   ['video/mp4; codecs="avc1.deadbeef"', 'audio/mp4; codecs="mp4a.40.E"'],
-                   'demuxed if URI');
-
-  delete master.mediaGroups.AUDIO.test.demuxed.uri;
-  assert.deepEqual(
-    mimeTypesForPlaylist_(master, media),
-    ['video/mp4; codecs="avc1.deadbeef, mp4a.40.E"', 'audio/mp4; codecs="mp4a.40.E"'],
-    'muxed if no URI and no playlists');
-
-  master.mediaGroups.AUDIO.test.demuxed.playlists = [{}];
-  assert.deepEqual(mimeTypesForPlaylist_(master, media),
-                   ['video/mp4; codecs="avc1.deadbeef"', 'audio/mp4; codecs="mp4a.40.E"'],
-                   'demuxed if no URI but playlists');
-});
-
-QUnit.test('uses audio codec from default group if not specified in media attributes',
-function(assert) {
-  const media = {
-    attributes: {
-      AUDIO: 'test',
-      CODECS: 'avc1.deadbeef'
-    },
-    segments: [
-      // signal fmp4
-      { map: 'test' }
-    ]
-  };
-  // dash audio playlist won't have a URI but will have resolved playlists
-  const master = {
-    mediaGroups: {
-      AUDIO: {
-        test: {
-          demuxed: {
-            default: true,
-            playlists: [{
-              attributes: {
-                CODECS: 'mp4a.40.E'
-              }
-            }]
-          }
-        }
-      }
-    },
-    playlists: [media]
-  };
-
-  assert.deepEqual(
-    mimeTypesForPlaylist_(master, media),
-    ['video/mp4; codecs="avc1.deadbeef"', 'audio/mp4; codecs="mp4a.40.E"'],
-    'uses audio codec from media group');
-
-  delete master.mediaGroups.AUDIO.test.demuxed.default;
-  assert.deepEqual(
-    mimeTypesForPlaylist_(master, media),
-    ['video/mp4; codecs="avc1.deadbeef"', 'audio/mp4; codecs="mp4a.40.2"'],
-    'uses default audio codec');
-});
-
-QUnit.module('Map Legacy AVC Codec');
-
-QUnit.test('maps legacy AVC codecs', function(assert) {
-  assert.equal(mapLegacyAvcCodecs_('avc1.deadbeef'),
-               'avc1.deadbeef',
-               'does nothing for non legacy pattern');
-  assert.equal(mapLegacyAvcCodecs_('avc1.dead.beef, mp4a.something'),
-               'avc1.dead.beef, mp4a.something',
-               'does nothing for non legacy pattern');
-  assert.equal(mapLegacyAvcCodecs_('avc1.dead.beef,mp4a.something'),
-               'avc1.dead.beef,mp4a.something',
-               'does nothing for non legacy pattern');
-  assert.equal(mapLegacyAvcCodecs_('mp4a.something,avc1.dead.beef'),
-               'mp4a.something,avc1.dead.beef',
-               'does nothing for non legacy pattern');
-  assert.equal(mapLegacyAvcCodecs_('mp4a.something, avc1.dead.beef'),
-               'mp4a.something, avc1.dead.beef',
-               'does nothing for non legacy pattern');
-  assert.equal(mapLegacyAvcCodecs_('avc1.42001e'),
-               'avc1.42001e',
-               'does nothing for non legacy pattern');
-  assert.equal(mapLegacyAvcCodecs_('avc1.4d0020,mp4a.40.2'),
-               'avc1.4d0020,mp4a.40.2',
-               'does nothing for non legacy pattern');
-  assert.equal(mapLegacyAvcCodecs_('mp4a.40.2,avc1.4d0020'),
-               'mp4a.40.2,avc1.4d0020',
-               'does nothing for non legacy pattern');
-  assert.equal(mapLegacyAvcCodecs_('mp4a.40.40'),
-               'mp4a.40.40',
-               'does nothing for non video codecs');
-
-  assert.equal(mapLegacyAvcCodecs_('avc1.66.30'),
-               'avc1.42001e',
-               'translates legacy video codec alone');
-  assert.equal(mapLegacyAvcCodecs_('avc1.66.30, mp4a.40.2'),
-               'avc1.42001e, mp4a.40.2',
-               'translates legacy video codec when paired with audio');
-  assert.equal(mapLegacyAvcCodecs_('mp4a.40.2, avc1.66.30'),
-               'mp4a.40.2, avc1.42001e',
-               'translates video codec when specified second');
 });
